Add field to set scan state based on warnings
Manual tests generate "warn" output. If a scan has no failures, it should by default
have state "Pass" even if it ran manual tests with "warn" output. Along with that, cluster scan
spec should have a field with which users can make the Scan fail if it ran tests with "warn"
output. This field will be an enum with values pass|fail and default of pass. It is an enum so in
the future if we decide to add a state specific to "warn" output we can extend this.