Add Warn field to additionalPrinterColumns, remove scoreWarning default
This commit adds the "Warn" field from status.summary as an additionalPrinterColumn.
It also removes the default value for scoreWarning. This is because the CRDs are of version
apiextensions.k8s.io/v1beta1. Setting a default in this version is not allowed. Although the
file crd.go specifies version as "v1", wrangler is generating CRD with version "apiextensions.k8s.io/v1beta1".
This could depend on the available versions in the cluster and can be investigated later.
For now in order to avoid any install issues, the default of "pass" is being removed. The logic to make the
scan fail checks the field specifically for "fail" value, so it won't be affected.